Update logic to validate email, update message when not verified, course name fixed
