+++ client/html/pages/panel/jobs/index.html
@@ -24,7 +24,8 @@
         </div>
         {{#unless currentUser.emails.[0].verified}}
           <section>
-            <h3>Você precisa confirmar o seu endereço de email antes de utilizar esse portal. Verifique seu email {{user.emails.[0].address}}</h3>
+            <h3>Você precisa confirmar o seu endereço de email antes de utilizar esse portal!</h3>
+            <h4>Verifique seu email {{user.emails.[0].address}} ou <a href="#" class="resend-verification-link">envie o link de verificação novamente</a></h4>
           </section>
         {{else}}
         <div class="card-body tab-content style-default-bright">
