+++ imports/startup/client/materialadmin-load.js
@@ -6,13 +6,7 @@ import {
 Template.HomeLayout.rendered = function() {
   this.autorun(_.bind(function() {
     Deps.afterFlush(function() {
-      // only initialize when DOM is ready
-      import '/imports/ui/materialadmin/core/source/AppNavigation.js';		
-      import '/imports/ui/materialadmin/core/source/AppOffcanvas.js';		
-      import '/imports/ui/materialadmin/core/source/AppCard.js';		
-      import '/imports/ui/materialadmin/core/source/AppForm.js';		
-      import '/imports/ui/materialadmin/core/source/AppNavSearch.js';		
-      import '/imports/ui/materialadmin/core/source/AppVendor.js';
+      // only initialize when DOM is ready		
       fixFloatingLabels();
     });
   }, this));
