+++ imports/ui/components/forms.js
@@ -8,6 +8,7 @@ Template.studentSignup.rendered = function() {
       // only initialize when DOM is ready
       setMasks();
       setCalendar(); 
+      setWizard();
       fixFloatingLabels();
       pwdValidate();
       autocompleteCEP();
@@ -171,5 +172,46 @@ function fixFloatingLabels() {
   });
   $('.floating-label .form-control').each(function() {
     $(this).after('<div class="form-control-line"></div>');
+  });
+}
+
+function setWizard() {
+  handleTabShow = function(tab, navigation, index, wizard) {
+    var total = navigation.find('li').length;
+    var current = index + 0;
+    var percent = (current / (total - 1)) * 100;
+    var percentWidth = 100 - (100 / total) + '%';
+
+    navigation.find('li').removeClass('done');
+    navigation.find('li.active').prevAll().addClass('done');
+
+    wizard.find('.progress-bar').css({
+      width: percent + '%'
+    });
+    $('.form-wizard-horizontal').find('.progress').css({
+      'width': percentWidth
+    });
+  };
+
+  var jqueryValidate = function(tab, navigation, index) {
+    var form = $('#wizard').find('.form-validation');
+    var valid = form.valid();
+    if (!valid) {
+      form.data('validator').focusInvalid();
+      return false;
+    }
+  }
+
+  $('#wizard').bootstrapWizard({
+    'nextSelector': '#next',
+    'previousSelector': '#previous',
+    'firstSelector': '#first',
+    'lastSelector': '#last',
+    'onTabShow': function(tab, navigation, index) {
+      handleTabShow(tab, navigation, index, $('#wizard'));
+    },
+    onNext: jqueryValidate,
+    onLast: jqueryValidate,
+    onTabClick: jqueryValidate
   });
 }
