+++ lib/router.js
@@ -187,3 +187,17 @@ FlowRouter.notFound = {
     BlazeLayout.render('PanelLayout', { header: "header", menubar: "menubar", main: "pageNotFound" });
   }
 };
+
+FlowRouter.route( '/verify-email/:token', {
+  name: 'verify-email',
+  action: function(params) {
+    Accounts.verifyEmail( params.token, ( error ) =>{
+      if ( error ) {
+        Meteor.call('displayErrors', error);
+      } else {
+        FlowRouter.go( '/' );
+        toastr.success('Você já pode acessar o painel', 'Email verificado!');
+      }
+    });
+  }
+});
